+++ SpringMVS-JDBS/src/main/java/com/example/springmvsjdbs/service/StudentService.java
@@ -7,11 +7,11 @@ import java.util.List;
 public interface StudentService {
     List<Student> findAll();
 
-    Student findById(Integer id);
+    Student findById(Long id);
 
     void update(Student student);
 
     void save(Student student);
 
-    void delete(Integer id);
+    void delete(Long id);
 }
